% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{extract_by_frequency}
\alias{extract_by_frequency}
\title{Extract and join all data frames with given frequency}
\usage{
extract_by_frequency(all_data, freq = 12)
}
\arguments{
\item{all_data}{all_data provided by `read_economic_indicator_files`}

\item{freq}{frequency to extract}
}
\value{
joined tibble of all series
}
\description{
Extract and join all data frames with given frequency
}
\details{
Extract and join all data frames with given frequency
}
\examples{
filenames <- "~/Documents/inflation_platon/data/Data/Economic Indicator_Russia Consumer Prices, Core CPI, Total, Index, DecPY=100_30 Nov 2017.xlsx"
# all_data <- read_economic_indicator_files(filenames)
# df_monthly <- extract_by_frequency(all_data, freq = 12)
}
