% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{extract_property}
\alias{extract_property}
\title{Extract cell with special text from vector}
\usage{
extract_property(first_col, property_text = "Interval: ")
}
\arguments{
\item{first_col}{character vector}

\item{property_text}{character}
}
\value{
extracted value
}
\description{
Extract cell with special text from vector
}
\details{
Extract cell with special text from vector
}
\examples{
first_col <- c("Sex: male", "Interval: A", "Group: B")
extract_property(first_col, "Group: ")
}
