% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{guess_frequency}
\alias{guess_frequency}
\title{Guess yearly frequency of dates vector}
\usage{
guess_frequency(timeline)
}
\arguments{
\item{timeline}{vector of dates}
}
\value{
integer frequency or NA
}
\description{
Guess yearly frequency of dates vector
}
\details{
Guess yearly frequency of dates vector
}
\examples{
data_vector <- as.Date(c("2017-03-31", "2017-04-30", "2017-05-31"))
guess_frequency(data_vector)
}
