% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation.project.R
\docType{package}
\name{inflation.project}
\alias{inflation.project}
\alias{inflation.project-package}
\title{inflation.project: project-package to forecast russian macro indicators with BVARs and alternative models}
\description{
Trying to test project-package idea.
}
\details{
Trying to test project-package idea...
}
