% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_economic_indicator}
\alias{read_economic_indicator}
\title{Read economic indicator file type}
\usage{
read_economic_indicator(filename)
}
\arguments{
\item{filename}{name of file}
}
\value{
list of two tibbles: with data and meta information
}
\description{
Read economic indicator file type
}
\details{
Read economic indicator file type
}
\examples{
filename <- "~/Documents/inflation_platon/data/Data/Economic Indicator_Russia Consumer Prices, Core CPI, Total, Index, DecPY=100_30 Nov 2017.xlsx"
# df_list <- read_economic_indicator(filename)
}
