% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_price_history}
\alias{read_price_history}
\title{Read price history file type}
\usage{
read_price_history(filename)
}
\arguments{
\item{filename}{name of file}
}
\value{
list of two tibbles: with data and meta information
}
\description{
Read price history file type
}
\details{
Read price history file type
}
\examples{
filename <- "~/Documents/inflation_platon/data/Data2/Price History [29].xlsx"
# two_df <- read_price_history(filename)
}
